fix(brain): kr-verifier $2类型歧义 + KR3/KR4改用progress-based采集 (#2022) (#2028)
* chore: [state] persist .dev-mode — Stage 1 Spec 完成

* fix(brain): kr-verifier $2类型歧义 + KR3/KR4改用progress-based采集

- kr-verifier.js: 修复 $2 参数类型歧义（PostgreSQL 无法同时推断 numeric+text）
  UPDATE 中 current_value = $2::numeric, ($2::numeric)::text
  消除 ZenithJoy KR1/KR2 verifier 的 'inconsistent types' 报错
- migration 225: 更新 KR3/KR4 verifier SQL
  从 status 映射伪进度（active=50%→25%）改为 okr_projects.progress 字段均值
  与实际开发状态对齐（当前 0%，随项目 progress 字段变化自动更新）


* docs: learning — kr-verifier $2 类型歧义 + status 伪进度陷阱

* chore(brain): bump EXPECTED_SCHEMA_VERSION 224→225

* docs(definition): schema_version 224→225

* test(brain): 更新 schema version 期望 224→225

---------

+++ packages/brain/src/kr-verifier.js
@@ -70,8 +70,8 @@ export async function runAllVerifiers() {
         await pool.query(`
           UPDATE key_results
           SET progress = $1,
-              current_value = $2,
-              metadata = COALESCE(metadata, '{}'::jsonb) || jsonb_build_object('metric_current', $2::text),
+              current_value = $2::numeric,
+              metadata = COALESCE(metadata, '{}'::jsonb) || jsonb_build_object('metric_current', ($2::numeric)::text),
               updated_at = NOW()
           WHERE id = $3
         `, [progress, currentValue, v.kr_id]);
